\name{PBD-internal}
\alias{pbd_loglik_choosepar}
\alias{pbd_loglik_rhs}
\alias{pbd_simplex}
\alias{brts2phylo}
\alias{pbd_bootstrap_sim}
\alias{pbd_incipient_rhs}
\alias{pbd_incipient}
\title{Internal PBD Functions}
\description{Internal PBD functions}
\details{These are not to be called by the user.}
\keyword{internal}
