<<<<<<< HEAD
\name{PBD-internal}
\alias{sampletree}
\alias{checkgood}
\alias{detphy}
\alias{L2phylo2}
\alias{pbd_reconstruct}
\alias{pbd_loglik_choosepar}
\alias{pbd_loglik_rhs}
\alias{pbd_simplex}
\alias{pbd_bootstrap_sim}
\alias{pbd_incipient_rhs}
\alias{pbd_incipient}
\title{Internal PBD Functions}
\description{Internal PBD functions}
\details{These are not to be called by the user.}
\keyword{internal}
=======
\name{PBD-internal}
\alias{sampletree}
\alias{checkgood}
\alias{detphy}
\alias{L2phylo2}
\alias{pbd_reconstruct}
\alias{pbd_loglik_choosepar}
\alias{pbd_loglik_rhs}
\alias{pbd_simplex}
\alias{pbd_bootstrap_sim}
\alias{pbd_incipient_rhs}
\alias{pbd_incipient}
\title{Internal PBD Functions}
\description{Internal PBD functions}
\details{These are not to be called by the user.}
\keyword{internal}
>>>>>>> 02bbac1e016a4e33fb6caad62db78bceaf1d65bc
