% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_durspec_mean.R
\name{pbd_durspec_mean_impl}
\alias{pbd_durspec_mean_impl}
\title{Actual calculation of the mean duration of speciation
(equations 19 and 20 of reference article)
assuming all inputs are correct}
\usage{
pbd_durspec_mean_impl(la2, la3, mu2)
}
\description{
Actual calculation of the mean duration of speciation
(equations 19 and 20 of reference article)
assuming all inputs are correct
}
\references{
Etienne, Rampal S., and James Rosindell. "Prolonging the past
  counteracts the pull of the present: protracted speciation can explain
  observed slowdowns in diversification." Systematic
  Biology 61.2 (2012): 204-213.
}
\seealso{
pbd_durspec_mean
}
\author{
Rampal S. Etienne
}
