% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_durspec_moment.R
\name{pbd_durspec_moment}
\alias{pbd_durspec_moment}
\title{Moments of duration of speciation under protracted birth-death model of
diversification}
\usage{
pbd_durspec_moment(pars, order)
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr \code{pars[1]} corresponds to b (=
la_3 in Etienne & Rosindell R2012) = speciation initiation rate \cr
\code{pars[2]} corresponds to la_1 (= la_2 in Etienne & Rosindell 2012) =
speciation completion rate \cr \code{pars[3]} corresponds to mu_2 (= mu_i in
ER2012) = extinction rate of incipient species \cr}

\item{order}{order of the moment to compute (1 is first moment, giving the
mean)}
}
\value{
The moment of the duration of speciation
}
\description{
pbd_durspec_moment computes the moments of the duration of speciation under
the protracted speciation model for a given set of parameters
}
\examples{
 pbd_durspec_moment(pars = c(0.5,0.3,0.1),2)
}
\seealso{
\code{\link{pbd_durspec_density}}\cr
\code{\link{pbd_durspec_cumdensity}}\cr \code{\link{pbd_durspec_mean}}\cr
\code{\link{pbd_durspec_mode}}\cr \code{\link{pbd_durspec_quantile}}\cr
\code{\link{pbd_durspec_var}}
}
\author{
Rampal S. Etienne
}
\keyword{models}
