% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_durspec_mean.R
\name{pbd_mean_durspec}
\alias{pbd_mean_durspec}
\title{Calculate the mean duration of speciation}
\usage{
pbd_mean_durspec(lambda_2, lambda_3, mu_2)
}
\arguments{
\item{lambda_2}{speciation completion rate, in probability per time unit}

\item{lambda_3}{speciation initiation rate of incipient species, in
probability per time unit}

\item{mu_2}{incipient species extinction rate, in probability per time unit}
}
\value{
the means duration of speciation, in time units
}
\description{
Calculate the mean duration of speciation
}
\examples{
  lambda_2 <- 0.1 # speciation completion rate
  lambda_3 <- 0.2 # speciation initiation rate of incipient species
  mu_2 <- 0.3 # incipient species extinction rate
  mean_durspec <- pbd_mean_durspec(lambda_2, lambda_3, mu_2)
  expected_mean_durspec <- 3.242955
  tolerance <- 0.0000001
  if (abs(mean_durspec - expected_mean_durspec) > tolerance) {
    stop("Duration of speciation must match expectation")
  }
}
\references{
Etienne, Rampal S., and James Rosindell. "Prolonging the past
  counteracts the pull of the present: protracted speciation can explain
  observed slowdowns in diversification." Systematic
  Biology 61.2 (2012): 204-213.
}
\seealso{
pbd_durspec_mean
}
