% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec.R
\name{pbd_numspec_mean}
\alias{pbd_numspec_mean}
\title{Mean number of species under protracted birth-death model of
diversification}
\usage{
pbd_numspec_mean(
  pars,
  parsf = c(function(t, pars) {
     pars[1]
 }, function(t, pars) {
     pars[2]
 },
    function(t, pars) {
     pars[3]
 }, function(t, pars) {
     pars[4]
 }),
  age,
  soc = 2,
  methode = "lsoda"
)
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr
\code{pars[1]} corresponds to b (=
la_1 = la_3 in Etienne & Rosindell R2012) = speciation initiation rate \cr
\code{pars[2]} corresponds to mu_1 (= mu_g in
ER2012) = extinction rate of good species \cr
\code{pars[3]} corresponds to la_1 (= la_2 in Etienne & Rosindell 2012) =
speciation completion rate \cr
\code{pars[4]} corresponds
to mu_2 (= mu_i in ER2012) = extinction rate of incipient species \cr}

\item{parsf}{Vector of functions how the rates depend on time, default
functions are constant functions of the parameters in pars1: \cr \cr
\code{parsf[1]} corresponds to time-dependence of b (= la_1 in Etienne &
Rosindell R2012) = speciation initiation rate \cr \code{parsf[2]}
corresponds to time-dependence of mu_1 (= mu_g in Etienne & Rosindell 2012)
= extinction rate of good species \cr \code{parsf[3]} corresponds to
tiem-dependence of la_1 (= la_2 in Etienne & Rosindell 2012) = speciation
completion rate \cr \code{parsf[4]} corresponds to time-dependence of mu_2
(= mu_i in ER2012) = extinction rate of incipient species \cr \cr}

\item{age}{the stem or crown age (see soc)}

\item{soc}{specify whether it is the stem or the crown age}

\item{methode}{Sets which method should be used in the ode-solver. Default
is 'lsoda'. See package deSolve for details.}
}
\value{
The expected number of representative species
}
\description{
pbd_numspec_mean computes the mean number of species under the
protracted speciation model for a given set of parameters
}
\details{
The number of representative species at time t = T when
starting with one lineage at time t = 0 is geometrically distributed with
parameter equal to g(T) where g(T) follows from the set of ODEs in Etienne
et al. 2014.
}
\examples{
pbd_numspec_mean(pars = c(0.3,0.1,0.5,0.1), age = 10, soc = 2)
}
\seealso{
\code{\link{pbd_numspec_quantile}}\cr
\code{\link{pbd_numspec_median}}
}
\author{
Rampal S. Etienne
}
\keyword{models}
