% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec_checked.R
\name{pbd_numspec_mean_checked}
\alias{pbd_numspec_mean_checked}
\title{Calcates the mean number of species under protracted birth-death model of
diversification. This is a checked version with a more explicit interface
than \code{\link{pbd_numspec_mean}}.}
\usage{
pbd_numspec_mean_checked(erg, eri, scr, sir, crown_age)
}
\arguments{
\item{erg}{extinction rate of a good species}

\item{eri}{extinction rate of an incipient species}

\item{scr}{speciation completion rate}

\item{sir}{speciation initiation rate. This function assumes the
speciation initiation rates of good and incipient species are equal.}

\item{crown_age}{the crown age}
}
\value{
The expected number of representative species
}
\description{
Calcates the mean number of species under protracted birth-death model of
diversification. This is a checked version with a more explicit interface
than \code{\link{pbd_numspec_mean}}.
}
\examples{
 mean_n_species <- pbd_numspec_mean_checked(
   erg = 0.12,
   eri = 0.23,
   scr = 0.34,
   sir = 0.45,
   crown_age = 0.56
 )
 expected_mean_n_species <- 1.046121595
 testthat::expect_equal(mean_n_species, expected_mean_n_species)
}
\seealso{
\code{\link{pbd_numspec_mean}} provides for setting a time-dependence
  in the parameters and/or specify a stem age.
  \code{\link{pbd_numspec_median_checked}} calculates the median number of species.
}
\author{
Richel J.C. Bilderbeek
}
