% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec_vec.R
\name{pbd_numspec_vec}
\alias{pbd_numspec_vec}
\title{Mean number of good and incipient species under protracted birth-death
model of diversification}
\usage{
pbd_numspec_vec(pars, age, initvec = c(1, 0))
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr
\code{pars[1]} corresponds to b_G (=
la_1 in Etienne & Rosindell R2012) = speciation initiation rate of good
species\cr
\code{pars[2]} corresponds to b_I (=
la_3 in Etienne & Rosindell R2012) = speciation initiation rate of incipient
species\cr
\code{pars[3]} corresponds to mu_G = extinction rate of good species \cr
\code{pars[4]} corresponds to mu_I = extinction rate of incipient species \cr
\code{pars[5]} corresponds to la (= la_2 in Etienne & Rosindell 2012) =
speciation completion rate \cr}

\item{age}{the stem age}

\item{initvec}{initial vector of good and incipient species. Default is 1 good
species.}
}
\value{
The expected number of good and incipient species
}
\description{
pbd_numspec_vec computes the mean number of good and incipient
species under the protracted speciation model for a given set of parameters
}
\details{
The result follows from a simple 2D system of differential equations
for G and I.
}
\examples{
pbd_numspec_vec(pars = c(0.3,0.3,0.1,0.1,0.2), age = 10)
}
\seealso{
\code{\link{pbd_numspec_mean}}\cr
\code{\link{pbd_numspec_quantile}}\cr
\code{\link{pbd_numspec_median}}
}
\author{
Rampal S. Etienne
}
\keyword{models}
