% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec_vec2.R
\name{pbd_numspec_vec2}
\alias{pbd_numspec_vec2}
\title{Mean number of good species with n incipient species and of n orphaned
incipient species under protracted birth-death model of diversification}
\usage{
pbd_numspec_vec2(
  pars = c(0.3, 0.3, 0.1, 0.1, 100),
  age = 10,
  initvec = c(1, rep(0, 20)),
  abstolint = 1e-10,
  reltolint = 1e-10,
  methode = "odeint::runge_kutta_cash_karp54"
)
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr
\code{pars[1]} corresponds to b_G (=
la_1 in Etienne & Rosindell R2012) = speciation initiation rate of good
species\cr
\code{pars[2]} corresponds to b_I (=
la_3 in Etienne & Rosindell R2012) = speciation initiation rate of incipient
species\cr
\code{pars[3]} corresponds to mu_G = extinction rate of good species \cr
\code{pars[4]} corresponds to mu_I = extinction rate of incipient species \cr
\code{pars[5]} corresponds to la (= la_2 in Etienne & Rosindell 2012) =
speciation completion rate \cr}

\item{age}{the stem age}

\item{initvec}{initial vector. Default is 1 good species with 0 incipient
species.}

\item{abstolint}{absolute tolerance in integration}

\item{reltolint}{relative tolerance in integration}

\item{methode}{integration method, one of the odeint solvers}
}
\value{
The expected number of good species with n incipient specis and the number of
caes with n orphaned incipient siblings
}
\description{
pbd_numspec_vec2 computes the mean number of good species with n
incipient daughter species and the number of cases with n orphaned incipient
siblings under the protracted speciation model for a given set of parameters
}
\examples{
pbd_numspec_vec2(pars = c(0.3,0.3,0.1,0.1,0.2), age = 10)
}
\seealso{
\code{\link{pbd_numspec_mean}}\cr
\code{\link{pbd_numspec_quantile}}\cr
\code{\link{pbd_numspec_median}}\cr
\code{\link{pbd_numspec_vec}}
}
\author{
Rampal S. Etienne
}
\keyword{models}
