% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_durspec_mean.R
\name{pbd_mean_durspec}
\alias{pbd_mean_durspec}
\title{Calculate the mean duration of speciation
(equations 19 and 20 of reference article)}
\usage{
pbd_mean_durspec(eri, scr, siri)
}
\arguments{
\item{eri}{extinction rate of the incipient species,
or mu_2 in article,
in probability per time unit}

\item{scr}{speciation completion rate,
or lambda_2 in article,
in probability per time unit}

\item{siri}{speciation initiation rate of incipient species, or
lambda_3 in article,
in probability per time unit}
}
\value{
the means duration of speciation, in time units
}
\description{
Calculate the mean duration of speciation
(equations 19 and 20 of reference article)
}
\examples{
  eri <- 0.1 # extinction rate of incipient species
  scr <- 0.2 # speciation completion rate
  siri <- 0.3 # speciation initiation rate of incipient species
  mean_durspec <- pbd_mean_durspec(eri, scr, siri)
  expected_mean_durspec <- 2.829762
  testthat::expect_equal(mean_durspec, expected_mean_durspec,
    tolerance = 0.000001)
}
\references{
Etienne, Rampal S., and James Rosindell. "Prolonging the past
  counteracts the pull of the present: protracted speciation can explain
  observed slowdowns in diversification." Systematic
  Biology 61.2 (2012): 204-213.
}
\seealso{
pbd_durspec_mean
}
