% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_ML.R
\name{pbd_ML}
\alias{pbd_ML}
\title{Likelihood maximization for protracted birth-death model of diversification}
\usage{
pbd_ML(brts, initparsopt = c(0.2, 0.1, 1),
  idparsopt = 1:length(initparsopt), idparsfix = NULL, parsfix = NULL,
  exteq = 1, parsfunc = c(function(t, pars) {     pars[1] }, function(t,
  pars) {     pars[2] }, function(t, pars) {     pars[3] }, function(t, pars) { 
     pars[4] }), missnumspec = 0, cond = 1, btorph = 1, soc = 2,
  methode = "lsoda", n_low = 0, n_up = 0, tol = c(1e-06, 1e-06, 1e-06),
  maxiter = 1000 * round((1.25)^length(idparsopt)), optimmethod = "subplex",
  verbose = TRUE)
}
\arguments{
\item{brts}{branching times (positive, from present to past).}

\item{initparsopt}{the initial parameter values:
initparsopt[1] = b (= la_1 in ER2012) = speciation initiation rate.
initparsopt[2] = mu_1 (= mu_g in ER2012) = extinction rate of good species.
initparsopt[3] = la_1 (= la_2 in ER2012) = speciation completion rate.
initparsopt[4] = mu_2 (= mu_i in ER2012) = extinction rate of incipient species.}

\item{exteq}{incipient species have the same (1) or different (0) extinction rate as good species}

\item{parsfunc}{functions of parameters}

\item{missnumspec}{number of missing species}

\item{cond}{conditioning:
cond = 0 conditioning on stem or clade age.
cond = 1 conditioning on age and non-extinction of the phylogeny.
cond = 2 conditioning on age and on number of extant taxa}

\item{btorph}{likelihood of branching times (0) or phylogeny (1), differ by a factor (S - 1)! where S is the number of extant species}

\item{soc}{stem (1) or crown (2) age}

\item{methode}{method of the numerical integration; see package deSolve for details}

\item{n_low}{lower bound on number of species (cond = 2)}

\item{n_up}{upper bound on number of species (cond = 2)}

\item{tol}{tolerance in optimization:}

\item{maxiter}{the maximum number of iterations in the optimization}

\item{optimmethod}{'subplex' (current default) or 'simplex' (default of previous versions)}

\item{verbose}{create output while performing algorithm}

\item{res}{resolution of the method; res should be larger than the total number of species}
}
\value{
Logarithm of the maximum likelihood estimate
}
\description{
Likelihood maximization for protracted birth-death model of diversification
}
\note{
'max(brts)' equals the crown age and
  'min(brts)' equals the most recent branching time
}
\examples{
  pbd_ML(1:10,initparsopt = c(0.2,0.01,0.3), exteq = 1)
}
\author{
Rampal S. Etienne
}
\seealso{
bdd_loglik
}

