% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbd_numspec.R
\name{pbd_numspec_median}
\alias{pbd_numspec_median}
\title{Median number of species under protracted birth-death model of
diversification}
\usage{
pbd_numspec_median(pars, parsf = c(function(t, pars) {     pars[1] },
  function(t, pars) {     pars[2] }, function(t, pars) {     pars[3] },
  function(t, pars) {     pars[4] }), age, soc = 2, methode = "lsoda")
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr
\code{pars[1]} corresponds to b (=
la_1 = la_3 in Etienne & Rosindell R2012) = speciation initiation rate \cr
\code{pars[2]} corresponds to mu_1 (= mu_g in
ER2012) = extinction rate of good species \cr
\code{pars[3]} corresponds to la_1 (= la_2 in Etienne & Rosindell 2012) =
speciation completion rate \cr
\code{pars[4]} corresponds
to mu_2 (= mu_i in ER2012) = extinction rate of incipient species \cr}

\item{age}{the stem or crown age (see soc)}

\item{soc}{specify whether it is the stem or the crown age}
}
\value{
The median number of representative species
}
\description{
pbd_numspec_median computes the median number of species under the
protracted speciation model for a given set of parameters
}
\examples{
pbd_numspec_median(pars = c(0.3,0.1,0.5,0.1), age = 10, soc = 2)
}
\seealso{
\code{\link{pbd_numspec_quantile}}\cr
\code{\link{pbd_numspec_mean}}
}
\author{
Rampal S. Etienne
}
\keyword{models}
